Strip absent (Nothing) fields.  All tests pass.
This was causing problems when creating bank accounts with IBAN but no
BIC.  MangoPay doesn't accept a null BIC, instead the BIC field needs to
be omitted altogether.

Although not necessary in order to make the tests pass, I've also
changed all other objects to strip absent fields.  This is in order to
be consistent across the whole library.